Introduce a new TestCase and move the generate_order() method into that class
