--- tests/SampleTest.php
@@ -1,20 +1,0 @@
-<?php
-/**
- * An example test.
- *
- * @package Nexcess\LimitOrders
- */
-
-namespace Tests;
-
-use WP_UnitTestCase as TestCase;
-
-class SampleTest extends TestCase {
-
-	/**
-	 * @test
-	 */
-	public function true_should_be_true() {
-		$this->assertTrue( true );
-	}
-}
